refactor: extract makeMetricsKey
